Huge HTML/CSS Overhaul, and behind-the-scenes work
Complete redesign of HTML/CSS and page aesthetics.
Completed a lot of behind-the-scenes improvements allowing the program
to work.
Completed parts of the profile page.
Completed Sign Up and Sign In pages.
Animations on all Webpages.
Added a favicon to all webpages.
Fixed plethora of bugs dealing with authentication and profile viewing.
Fixed bugs regarding I/O with the database.