Class / Overview Color Changes & Set Current Date in Add Assignment Section
Changes made to class colors as well as overview box colors in order to
compliment each other more and match other colors throughout the
website.

Also, a small script was added to set the current date when adding an
assignment to make it slightly quicker for the user to select an
upcoming due date.