Changed sorting method and tests
Changed to a bubble sort, it performed faster in all tests.
Changed the way we calculate the number of days in a date.
Simplified some code in Parent, Student, and Teacher
Changed all uses of the method ref to find.where().eq()
Fixed an issue where the default color would appear on the assignment
boxes when it wasn't supposed to
Added trailing slashes to all routes
Starting writing Unit tests