Security checks and fixed download
Added a security check to every function that makes sure the user who is
requesting has the correct permissions to perform that function.
Trying a new thing for the download feature to get it to work.
Fixed delete accounts on the parent account.
Create unauthorizedError.scala.html for when a user attempts to access a
function that they don't have access to.
Fixed word colors on the edit school class page.
Fixed an issue where it would say the password wasn't long enough during
account editing when they hadn't even put in a password.
Fixed an issue where 1 would be an invalid grade level when creating a
child.
Fixed an issue where editing a child account would completely fail if
one part failed.
Enabled parents to delete a child account and be redirected back to the
parent account rather than the index.
Fixed some wording in teacherProfile.scala.html
Fixed some things in tests to reflect new password requirements.