Bug fixes
Fixed delete for parent.
Made it so that after deleting a schoolClass on teacher profile, it
would default to the Classes section rather than Overview.
Made it so that after signing up for a teacher or parent account, the
Overview tab would light up blue like normal.
Added more descriptive messages on parent profile for when they have no
children or have children with no homework.
Made it so that when using the default color on student and teacher
accounts, it wouldn't show the default red color as the class color but
rather default to the color of the box.
Removed how it showed the class password on teacher accounts in the
Classes tab.