Mobile Improvements
-Removed hamburger icon animation into the X. Didn’t appear on iPhone
and ultimately decided it would be best to keep it simple like Matthew
initially had.
-Fixed spacing issues on sign up pages and parent account settings
-Fixed mobile viewing on the contact page
-The images that slid in on the home page were not showing at all on
iPhone, so I made it one complete image and made an iPhone breakpoint
for that to appear on mobile devices.
-Removed the earthquaking shake on unauthorized page, no more seizures
haha.
-Adjusted some content inside of overview boxes in profile for mobile
devices.
-Removed tooltips on mobile devices. I found that it was annoying you
would have to click the icon twice for the action to occur. Plus the
icons give clear enough direction as to what they will do in my opinion.