Multiple concurrent sessions in browser
Enabled it so that multiple accounts can be opened on the same browser
without the "unauthorized" error. This is a big change for parent
accounts because before, it would essentially log the parent out.
On the UnauthorizedError page, fixed an anchor tag so it wouldn't
underline a space.
When a parent clicks on a child account, it will open in a new tab.
When a student finishes an assignment, it will no longer set the active
tab to the finished assignments tab, but instead the overview tab.
Added a trailing slash to one of the routes