Bug fixes with multiple cookies
Made it so that logout will no longer clear the entire session, but
instead just set the value of the key of the user to "Logged Out".
Added comments about the variables to Students.java.
Added comments to the methods in Utilities.java.
Removed code that was for testing in Utilities.java.
Added trim() to methods that were using string comparison in Utilities.
Created a method in Utilities that would take in a value and find that
value in the session and return the index of it.
Removed an unused import in Utilities.java.
Added the parent UUID argument to the logout href on
parentProfile.scala.html.
Added the student UUID argument to the logout href on
studentProfile.scala.html.
Added the teacher UUID argument to the logout href on the
teacherProfile.scala.html.
Added the conf\application.conf file to .gitignore so no one can
accidentally push it (because it has a password to their MySQL DB).
Added the UUID parameter to logout in routes.
Attempted fixes in ApplicationTest.java to fix tests.